Updates to Call Center Lead Mission
I made some updates, calling out specifically the focus to build up the capacity of the White House Hotline to serve as our call center. This is going to be a process and will probably take the bulk of work if we ever want to get close to what we had with the HRC hotline. I would say that is the primary and most important objective for the call center lead as the White House Hotline will be the only call center serving VA.gov issues moving forward.