Adds initial holiday release freeze schedule.
Initial proposal for dates for holiday release suspensions.

Note that I'm proposing a few days around Thanksgiving, as well as an extended two-week freeze over the holidays. I'm open to pushback here but I really think we should be avoiding while most people are out, and set the expectation to others not to be expecting to release functionality over the holidays.