Formatting and organization updates
My edits mostly fell into the following areas:
- using parallel structure in headings and bullets
- using active voice
- trimming words to make sentences shorter
- using sentence case
- applying style guide recommendations for formatting

I removed the quick reference checklist in section about setting up the backend. If you want to add it back, I would format it as a checklist and not include the "Broker matrix and tagging" information since that's not really a step.