Update internal-tools doc to clarify that socks.sh and SwitchyOmega are alternatives (#14627)
Update socks.sh to avoid the use of *gov* which ended up causing issues with the script;
the problem is, it would try to load public keys as well, and fail. Thus, it
has been changed to reference id_rsa_vagov, which is the suggested keyname.
It is also possible to override the keyname, and instructions are given
in the script for these cases.