Updated KPIs and added campaign link URLs
Updated the KPI table with feature user and specific events that can be monitored on the Mobile dashboard where applicable. Also updated the table with the App Store Analytics reports that will be used for other KPIs. 

Added a section explaining how to navigate to the data points in monthly dashboard below the table.

Added the campaign URLs that must be used in promo banners to refer users to download the app in order for App Store analytics to track impressions, downloads, and engagements by new users from the banners.