Fix Windows incompatibilities in file paths, other minor fixes
VA repositories should be cloneable on all platforms, so directory and
file names that are incompatible with some platforms should not be
used. Specifically, this commit fixes the following incompatibilities
with Windows:

* names containing question marks
* names containing colons
* names containing pipe symbols (`|`)
* names ending with space
* names ending with period
* file paths longer than 256 characters

In addition, though it isn't an actual incompatibility, file names
beginning with space should also be avoided, so a few of these are
also fixed in this commit.

In addition, this commit:

* deletes a superfluous file named ` .pdf` that was also committed
  under a different, correct name;
* fixes a typo I noticed in a file name while doing all the other
  cleanup; and
* fixes an invalid link in a MarkDown file that I happened to notice
  while doing all the other cleanup.