Create how-we-audit-2024.md
INITIAL COPY/PASTE: Moving version from shared support to CAIA folder, to avoid confusion. This is linked in the new accessibility Platform page, and may confuse users, as it says to contact shared support. Revising with CAIA contact information on next save.  This is a direct copy/paste for this version. Version Credit: DK and Shared Support Team