Updates to ADE Github Terms and Definitions
Updated the table description to describe how the information in the table is sorted from a high-level to a more granular level. Edits to the formatting of the table, breaking out bulleted information, pulling out key terms "checked task" and "new issues" for example. Added a linked footnote, to reduce the content of the table for "sub-issues."