Anchor links added to ticket types with footer note
Re: Anchor links and collapsed sections: Anchor links will not activate on a collapsed details section, unless it is expanded on the page. In this example, be sure to expand the collapsed details section "Expand to view descriptions of the three main ticket types" to activate the anchor links above it.