Added and updated template access tip
Made note that the repository is the source of truth. Added note to search on the term ADE, and note about internal flags and descriptions. Added footnote number 3, to alert new team members of ADE about linking to a relative URL with piping of assignees going stale and instead to link to the actual template, and provided an example.