Update request-contact-center-review.md
We have added a step (No. 4 under Step 1: Create the Product Guide) that requires project managers to include solution or workaround information for known errors. If the solution isn't currently known, we ask that project managers include the questions that call center agents should ask users so that the product team can begin researching the matter.